requirements.txt
Bootstrap_Flask==2.3.3
Flask_WTF==1.2.1
WTForms==3.0.1
Werkzeug~=2.3.8
Flask~=2.3.3
asgiref~=3.7.2
blinker~=1.6.2
click~=8.1.5
colorama~=0.4.6
colormap~=1.0.4
contourpy~=1.1.0
convcolors~=2.2.0
cycler~=0.11.0
Django~=4.2.3
dominate~=2.8.0
extcolors~=1.0.0
fonttools~=4.43.0
image~=1.5.33
itsdangerous~=2.1.2
Jinja2~=3.1.2
joblib~=1.3.1
kiwisolver~=1.4.4
MarkupSafe~=2.1.3
matplotlib~=3.7.2
numpy~=1.26.4
packaging~=23.1
pandas~=2.0.3
Pillow~=10.0.0
pyparsing~=3.0.9
python-dateutil~=2.8.2
pytz~=2023.3
scikit-learn~=1.3.0
scipy~=1.11.1
six~=1.16.0
sqlparse~=0.4.4
threadpoolctl~=3.2.0
tzdata~=2023.3
visitor~=0.1.3
python-dotenv~=1.0.1